fix(core): account for `{ Ref }` incompatibility between schema and CFN
This is the change in CDK that consumes the spec update https://github.com/cdklabs/awscdk-service-spec/pull/2328.

This accounts for `{ Ref }` not always returning what the schema says
that the primary identifier of a resource is. This is not unexpected or
wrong: the Registry Schema describes the behavior of Cloud Control API,
and the identifier uniquely identifies a resource in an account. We do
need to account for the differences though.

What we will do is:

- Retain the CC-API fields in the reference interface if we hvae the
  values.
- Fall back to the (smaller) CFN identifier fields if we don't have the
  values.
- Make the CFN identifier fields win if the identifiers are equal
  size (in which case we probably a Name vs ARN distinction).