fix(cloudwatch): skip MathExpression validation when prop is a token (#36487)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

`MathExpression` class has multiple props that needs to be alphanumeric, that check for this is already there, but the problem is the checker doesn't check first if it's token or not. so for any tokens it always throw an error since it has `$` even if the resolve of this token is alphanumeric string.

### Description of changes

I'm checking if it's token or not, and if it's i'm passing the alphanumeric checker, since in this case u can't check it.

### Describe any new or updated permissions being added




### Description of how you validated changes



### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*